actualizar scripts

+++ install.sh
@@ -1,3 +1,5 @@
-rm ~/.local/bin/today
+if [ -e "~/.local/bin/today" ]; then
+  rm ~/.local/bin/today
+fi
 
 cp today.sh ~/.local/bin/today
